modularize Colors
